Reconcile tablet container and lore tablet container
