+++ ItemChanger.Silksong/Containers/TabletContainer.cs
@@ -1,11 +1,12 @@
 ﻿using ItemChanger.Containers;
+using ItemChanger.Extensions;
+using ItemChanger.Silksong.Assets;
 using ItemChanger.Silksong.Components;
+using ItemChanger.Silksong.Util;
 using MonoDetour.DetourTypes;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using Unity.Burst.CompilerServices;
+using TeamCherry.Localization;
 using UnityEngine;
+using UnityEngine.SceneManagement;
 
 namespace ItemChanger.Silksong.Containers;
 
@@ -79,5 +80,29 @@ public class TabletContainer : Container
         
         item.Get();
         return ReturnFlow.SkipOriginal;
+    }
+
+    /// <summary>
+    /// Instantiates a Weaver lore tablet (like the one found in Weaver_08).
+    /// Its text is determined by invoking the message provider each time the tablet is read.
+    /// </summary>
+    /// <returns>
+    /// The spawned lore tablet.
+    /// </returns>
+    public static GameObject InstantiateWeaverTablet(Scene scene, Func<string> messageProvider)
+    {
+        GameObject tablet = GameObjectKeys.LORE_TABLET_WEAVER.InstantiateAsset(scene);
+
+        string modKey = "IC_WEAVER_TABLET";
+        LocalisedString s = new(Localization.Sheet, modKey);
+        BasicNPC npc = tablet.FindChild("Inspect Region (1)")!.GetComponent<BasicNPC>();
+        npc.StartingDialogue += () =>
+        {
+            Language._currentEntrySheets[Localization.Sheet][modKey] = messageProvider();
+        };
+        npc.talkText = [s];
+        npc.repeatText = s;
+        npc.returnText = s;
+        return tablet;
     }
 }
